Default to "west flash" when flashing with zephyr
There is no real reason to check for platform validity there, and as we
use zephyr this seems like a sane default. This also allows the usage of
any zephyr-supported board (which can actually run uros) by just adding
its name to the supported_platforms file (previously, a special casing
in flash.sh was also needed).

Signed-off-by: Robert Wilbrandt <robert@stamm-wilbrandt.de>