Improve flash step for zephyr RTOS (#225)
* Clean up zephyr flash step

There were some issues here:
- The handling of the nucleo_h743zi platform seems to have been broken
  (caused by a wrong nesting in the platform selection)
- Not all cases checked for the build artifact before using it
- Console output differed between different platforms
- Indentation was inconsistent

Signed-off-by: Robert Wilbrandt <robert@stamm-wilbrandt.de>

* Unify zephyr flash openocd handling

This changes the default behavior of the olimex-stm32-e407 platform,
which previously didn't run directly after flashing. I could not find a
reason why this should be the case and thus made it consistent with the
"west flash" and nucleo_h743zi behavior.

Signed-off-by: Robert Wilbrandt <robert@stamm-wilbrandt.de>

* Default to "west flash" when flashing with zephyr

There is no real reason to check for platform validity there, and as we
use zephyr this seems like a sane default. This also allows the usage of
any zephyr-supported board (which can actually run uros) by just adding
its name to the supported_platforms file (previously, a special casing
in flash.sh was also needed).

Signed-off-by: Robert Wilbrandt <robert@stamm-wilbrandt.de>